refactor: relocated store
