+++ src/app/providers/ProductsProvider/ProductsProvider.tsx
@@ -1,6 +1,6 @@
 import type { JSX } from 'react'
 import { createContext, type ReactNode, useState } from 'react'
-import { type ProductsStateType, store } from '@/app/store/AppStore'
+import { productsState, type ProductsStateType } from './ProductsState'
 
 export type IProductsContext = {
   products: ProductsStateType
@@ -8,12 +8,12 @@ export type IProductsContext = {
 }
 
 export const ProductsContext = createContext<IProductsContext>({
-  products: store.products,
+  products: productsState,
   setProducts: (): void => {},
 })
 
 export function ProductsProvider({ children }: { children: ReactNode }): JSX.Element {
-  const [products, setProductsState] = useState<ProductsStateType>(store.products)
+  const [products, setProductsState] = useState<ProductsStateType>(productsState)
 
   const setProducts = (newProducts: ProductsStateType): void => {
     setProductsState(newProducts)
