+++ src/app/providers/ProductsProvider/ProductsState.ts
@@ -1,0 +1,7 @@
+export type ProductsStateType = {
+  someProducts: []
+}
+
+export const productsState: ProductsStateType = {
+  someProducts: [],
+}
