+++ src/app/providers/UserProvider/UserProvider.tsx
@@ -1,5 +1,5 @@
-import { type UserStateType, store } from '@/app/store/AppStore'
 import type { JSX } from 'react'
+import { userState, type UserStateType } from './UserState'
 import { createContext, type ReactNode, useState } from 'react'
 
 export type IUserContext = {
@@ -8,12 +8,12 @@ export type IUserContext = {
 }
 
 export const UserContext = createContext<IUserContext>({
-  user: store.user,
+  user: userState,
   setUser: (): void => {},
 })
 
 export function UserProvider({ children }: { children: ReactNode }): JSX.Element {
-  const [user, setUserState] = useState<UserStateType>(store.user)
+  const [user, setUserState] = useState<UserStateType>(userState)
 
   const setUser = (newUser: UserStateType): void => {
     setUserState(newUser)
